requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer
Django
djangorestframework==3.16.1
idna==3.10
methodism==0.3.4
pillow==11.3.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
django-environ==0.12.0
PyJWT==2.10.1
